make x64 happy

+++ trunk/src/MurmurHash/MurmurHashAligned2.cpp
@@ -1,3 +1,5 @@
+#include <stddef.h>
+
 //-----------------------------------------------------------------------------
 // MurmurHashAligned2, by Austin Appleby
 
@@ -17,7 +19,7 @@ unsigned int MurmurHashAligned2 ( const void * key, int len, unsigned int seed )
 
 	unsigned int h = seed ^ len;
 
-	int align = (int)data & 3;
+	size_t align = (size_t)data & 3;
 
 	if(align && (len >= 4))
 	{
